Handle owner and stack when resolving partially deduped elements
This is a follow-up from #30528 to not only handle props (the critical
change), but also the owner and stack of a referenced element.

Handling stacks here is rather academic because the Flight Server
currently does not deduplicate owner stacks. And if they are really
identical, we should probably just dedupe the whole element.